requirements.txt
bluequbit>=0.18.0
qiskit>=1.0.0
pyqubo>=1.4.0
numpy>=1.24.0
flask>=3.0.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
